requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
colorama==0.4.3
Django==3.1.1
django-bootstrap4==2.2.0
django-cors-headers==3.5.0
django-tenants==3.2.1
djangorestframework==3.11.1
importlib-metadata==1.7.0
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.6.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0
